refactor(tracestore): Make SearchDepth and PageSize unsigned on internal query types
Closes #9637

+++ cmd/jaeger/internal/extension/jaegerquery/internal/grpc_handler_test.go
@@ -978,7 +978,7 @@ func TestFindTracesRefusedByQueryService_GRPC(t *testing.T) {
 				ServiceName:  "service",
 				StartTimeMin: time.Now().Add(-10 * time.Minute),
 				StartTimeMax: time.Now(),
-				SearchDepth:  tracestore.MaxSearchDepth + 1,
+				SearchDepth:  int32(tracestore.MaxSearchDepth + 1),
 			},
 			wantErr: "search depth must be in [0, 10000]",
 		},
